+++ pages/project-details/[id]/view-details/[viewId].js
@@ -1,9 +1,8 @@
 import { useRouter } from "next/router";
 import styled from "styled-components";
-import Image from "next/image";
 import Link from "next/link";
-import ModalDeleteButton from "@/components/ModalDeleteButton";
 import { useState } from "react";
+import DetailsHeader from "@/components/Header";
 
 export default function ViewDetails({ views, setViews }) {
   const router = useRouter();
@@ -25,28 +24,13 @@ export default function ViewDetails({ views, setViews }) {
 
   return (
     <>
-      <Header>
-        <TitleWrapper>
-          <Title>{name}</Title>
-          {sketch ? (
-            <StyledImage
-              src={sketch}
-              alt={`here should be a sketch of your view`}
-              width="100"
-              height="100"
-            />
-          ) : (
-            <NoSketchText>no sketch here</NoSketchText>
-          )}
-        </TitleWrapper>
-      </Header>
+      <DetailsHeader
+        name={name}
+        sketch={sketch}
+        setEntries={setViews}
+        currentId={currentView.id}
+      />
       <Main>
-        <ModalDeleteButton
-          showModalDelete={showModalDelete}
-          setShowModalDelete={setShowModalDelete}
-          currentProject={currentView}
-          setEntries={setViews}
-        />
         {description ? (
           <DescriptionSection>
             <Subtitle>Description</Subtitle>
@@ -62,10 +46,6 @@ export default function ViewDetails({ views, setViews }) {
   );
 }
 
-const StyledImage = styled(Image)`
-  object-fit: cover;
-`;
-
 const DescriptionText = styled.p`
   overflow: scroll;
   overflow-wrap: break-word;
@@ -76,36 +56,8 @@ const DescriptionSection = styled.section`
   height: 20vh;
 `;
 
-const Header = styled.header`
-  width: 100vw;
-  padding-top: 20px;
-`;
-
-const Title = styled.h1`
-  margin: 10px;
-  overflow-wrap: break-word;
-  overflow: hidden;
-  font-size: 1.5em;
-  width: 50vw;
-`;
-
 const Main = styled.main`
   margin: 0 10px;
-`;
-
-const NoSketchText = styled.p`
-  border: 1px solid black;
-  width: 100px;
-  height: 100px;
-  padding: 5px;
-  text-align: center;
-`;
-
-const TitleWrapper = styled.div`
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: 0 20px;
 `;
 
 const Subtitle = styled.h2`
